refactor: consolidate type definitions into centralized types module
- Move hunt-related types (HuntStatus, StoredHunt, HuntInfo) from huntStore.ts to lib/types.ts
- Move clue-related types (ClueInfo, Clue, ClueRow) from hunt.ts to lib/types.ts
- Move activity feed types (ActivityEventType, ActivityEvent) from activityFeed.ts to lib/types.ts
- Move player registration types (PlayerProgress, RegistrationStatus, RegistrationResult) from player-registration.ts to lib/types.ts
- Move transaction result types (CreateHuntResult, SubmitAnswerResult, ActivateHuntResult, AddClueResult) from hunt.ts to lib/types.ts
- Move leaderboard types (LeaderboardEntry, LeaderboardDisplayEntry) from hunt.ts to lib/types.ts
- Update all contract and store modules to re-export types from centralized lib/types.ts
- Improves maintainability by establishing single source of truth for type definitions
