+++ app/dashboard/page.tsx
@@ -6,7 +6,8 @@ import { ArrowLeft } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { Header } from "@/components/Header"
 import { HuntDashboard } from "@/components/HuntDashboard"
-import { getCreatorHunts, updateHuntStatus, saveClueLocally, type StoredHunt } from "@/lib/huntStore"
+import type { StoredHunt } from "@/lib/types"
+import { getCreatorHunts, updateHuntStatus, saveClueLocally } from "@/lib/huntStore"
 import { activateHunt, addClue } from "@/lib/contracts/hunt"
 import { withTransactionToast } from "@/lib/txToast"
