+++ app/hunt/[id]/share.tsx
@@ -4,7 +4,8 @@ import { HuntControls } from "@/components/HuntControls";
 import { Button } from "@/components/ui/button";
 import { QrCode } from "lucide-react";
 import { QrCodeModal } from "@/components/QrCodeModal";
-import { StoredHunt, updateHuntStatus } from "@/lib/huntStore";
+import type { StoredHunt } from "@/lib/types";
+import { updateHuntStatus } from "@/lib/huntStore";
 import { useRouter } from "next/navigation";
 import { useState, useEffect } from "react";
 import { RegistrationButton } from "@/components/RegistrationButton";
