+++ components/RewardsPanel.tsx
@@ -4,19 +4,10 @@ import { Button } from "@/components/ui/button"
 import { Plus, Minus } from "lucide-react"
 import Trash from "@/components/icons/trash"
 import Coin from "@/components/icons/Coin"
-import { ReactNode, useState } from "react"
+import { useState } from "react"
+import type { Reward, RewardPlayerProgress } from "@/lib/types"
 
-export interface Reward {
-  place: number
-  amount: number
-  icon?: ReactNode
-}
-
-export interface PlayerProgress {
-  is_completed: boolean;
-  reward_claimed: boolean;
-  hunt_id?: number | string;
-}
+export type { Reward, RewardPlayerProgress as PlayerProgress }
 
 export interface RewardsPanelProps {
   rewards: Reward[];
@@ -24,7 +15,7 @@ export interface RewardsPanelProps {
   onAddReward?: () => void;
   onDeleteReward?: (place: number) => void;
   error?: string;
-  playerProgress?: PlayerProgress;
+  playerProgress?: RewardPlayerProgress;
   onClaimReward?: (hunt_id?: number | string) => Promise<void>;
 }
